Endpoints added
There are several endpoints added to this version:
get_all_items where we can obtain all the items in the system.
create_new_item where we can add a new item to the system.
edit_item_availability where we can specify if the item is currently available at the warehouse or not.

get_all_shipments where we can obtain all the shipments in the system.
create_new_shipment where we can add a new shipment to the system.
delete_a_shipment where we can delete a shipment.